yay bug fixs kinda have things working, player moves to gold if it can see it, it then gets stuck:
Agent:
got rid of the chain of changeDirection with hard coded directions because.... its predetermined what its equal to

GetGold:
was iterating over items on player inventory instead of items on map.
was returning startpoint instead of null when gold not found for debugging purposes before removed it.
heuristic uses absolute val when calculating manhatten distance

Graph:
yes im an idiot...
forgot to update direction to direction of current state.
got directions for moving east and west mixed up

State:
holy crap im an idiot...
Forgot to call the method that calculates and sets the heuristic and f cost...
and i forgot to update the gcost.
added methods to get direction, get the g cost, and print the path taken